Fixed Paddle Bug
Fixed the AI paddle bug where it wasn't moving after the 1st round ended.

Root cause ended up being due to the next round starting prior to destroying the ball from the previous round, which cause the array of balls to hold a reference to the previous rounds ball which was then destroyed.